adds context folder in hooks folder and moves all the contexts to it
