+++ frontend/src/components/authentication/sign-in/index.tsx
@@ -9,7 +9,7 @@ import {
 
 import { getFormData } from "../utils";
 import SignInForm from "./SigninForm";
-import useUserAuthContext from "../../../hooks/useUserAuthContext";
+import useUserAuthContext from "../../../hooks/context/useUserAuthContext";
 
 export default function SignIn() {
     const navigate = useNavigate();
