+++ frontend/src/components/cart-page/CartItemsSection.tsx
@@ -5,7 +5,7 @@ import { cartCardStyle } from "./cartStyles";
 import ItemsList from "../checkout/place-order/order-itmes/ItemsList";
 import EmptyCartMessage from "./EmptyCartMessage";
 import { HOME_PATH } from "../constants/path";
-import useShoppingCartContext from "../../hooks/useShoppingCartContext";
+import useShoppingCartContext from "../../hooks/context/useShoppingCartContext";
 
 
 const CartItemsSection = () => {
