+++ frontend/src/components/checkout/place-order/PaymentMethod.tsx
@@ -2,7 +2,7 @@ import { Card, CardActions, CardContent, Button, Typography } from '@mui/materia
 
 import { useNavigate } from 'react-router-dom';
 import { placeOrderComponentsStyles } from '../styles';
-import useCheckoutContext from '../../../hooks/useCheckoutContext';
+import useCheckoutContext from '../../../hooks/context/useCheckoutContext';
 
 const { container, title } = placeOrderComponentsStyles
