+++ frontend/src/components/navbar/UserSettingsMenu.tsx
@@ -2,9 +2,9 @@ import { Menu, MenuItem } from "@mui/material";
 import { ORDERS_HISTORY_PATH, UPDATE_USER_PATH } from "../constants/path";
 import { MY_ORDERS_TEXT, PROFILE_TEXT, SIGNOUT_TEXT } from "../constants/text";
 import { useNavigate } from "react-router-dom";
-import useAppContext from "../../hooks/useAppContext";
-import useUserAuthContext from "../../hooks/useUserAuthContext";
-import useMenuSettingsContext from "../../hooks/useMenuSettingsContext";
+import useAppContext from "../../hooks/context/useAppContext";
+import useUserAuthContext from "../../hooks/context/useUserAuthContext";
+import useMenuSettingsContext from "../../hooks/context/useMenuSettingsContext";
 
 type UserSettingsMenuProps = {
     isMenuOpen: boolean;
@@ -13,9 +13,8 @@ const UserSettingsMenu: React.FC<UserSettingsMenuProps> = ({ isMenuOpen }) => {
     const navigate = useNavigate();
     const { getMenuClickHandler } =
         useAppContext();
-    const { userOptionsOpen } = useMenuSettingsContext()
+    const { userOptionsOpen, handleMenuClose } = useMenuSettingsContext()
     const { handleSignOut } = useUserAuthContext();
-    const { handleMenuClose } = useMenuSettingsContext()
     // When isMenuOpen state is true this function return a dropdown menu that show some navigation options
     // to the user such as my Orders, Orders History etc...
     return (
