+++ frontend/src/components/navbar/UserSettingsMobileButton.tsx
@@ -1,8 +1,8 @@
 
 import { IconButton } from '@mui/material'
 import MoreIcon from "@mui/icons-material/MoreVert";
+import useMenuSettingsMobileContext from '../../hooks/context/useMenuSettingsMobileContext';
 
-import useMenuSettingsMobileContext from '../../hooks/useMenuSettingsMobileContext';
 
 const UserSettingsMobileButton = () => {
