+++ frontend/src/contexts/user-auth-context/index.tsx
@@ -6,10 +6,10 @@ import { initialShippingAddressData, userData } from "../app-context/data";
 import { NavigateFunction } from "react-router-dom";
 import { SIGNIN_PATH } from "../../components/constants/path";
 
-import useShoppingCartContext from "../../hooks/useShoppingCartContext";
-import useCheckoutContext from "../../hooks/useCheckoutContext";
-import useMenuSettingsMobileContext from "../../hooks/useMenuSettingsMobileContext";
-import useMenuSettingsContext from "../../hooks/useMenuSettingsContext";
+import useShoppingCartContext from "../../hooks/context/useShoppingCartContext";
+import useCheckoutContext from "../../hooks/context/useCheckoutContext";
+import useMenuSettingsMobileContext from "../../hooks/context/useMenuSettingsMobileContext";
+import useMenuSettingsContext from "../../hooks/context/useMenuSettingsContext";
 
 export const UserAuthContext = createContext({} as UserAuth);
