+++ frontend/src/hooks/context/useOrdersContext.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { OrdersContext } from "../contexts/orders-context";
+import { OrdersContext } from "../../contexts/orders-context";
 
 const useOrdersContext = () => {
     const { orderData, setOrderData } = useContext(OrdersContext);
