+++ frontend/src/hooks/context/useUserAuthContext.tsx
@@ -1,5 +1,5 @@
 import { useContext } from "react";
-import { UserAuthContext } from "../contexts/user-auth-context";
+import { UserAuthContext } from "../../contexts/user-auth-context";
 
 const useUserAuthContext = () => {
     const { userSignedIn, setUserSignedIn, handleSignOut } =
